+++ kie-dmn/kie-dmn-feel/src/main/java/org/kie/dmn/feel/runtime/functions/CustomFEELFunction.java
@@ -16,81 +16,20 @@
 
 package org.kie.dmn.feel.runtime.functions;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 
-import org.kie.dmn.api.feel.runtime.events.FEELEvent;
-import org.kie.dmn.api.feel.runtime.events.FEELEvent.Severity;
 import org.kie.dmn.feel.lang.EvaluationContext;
 import org.kie.dmn.feel.lang.ast.BaseNode;
-import org.kie.dmn.feel.runtime.events.FEELEventBase;
-import org.kie.dmn.feel.runtime.events.InvalidInputEvent;
-import org.kie.dmn.feel.runtime.events.InvalidParametersEvent;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class CustomFEELFunction extends BaseFEELFunction {
+public class CustomFEELFunction extends AbstractCustomFEELFunction<BaseNode> {
 
-    private static final Logger logger = LoggerFactory.getLogger( CustomFEELFunction.class );
-
-    private final List<BaseFEELFunction.Param> parameters;
-    private final BaseNode     body;
-
-    public CustomFEELFunction(String name, List<BaseFEELFunction.Param> parameters, BaseNode body) {
-        super( name );
-        this.parameters = parameters;
-        this.body = body;
-    }
-
-    public List<List<String>> getParameterNames() {
-        return Arrays.asList(parameters.stream().map(BaseFEELFunction.Param::getName).collect(Collectors.toList()));
-    }
-
-    public FEELFnResult<Object> invoke(EvaluationContext ctx, Object[] params ) {
-        if( params.length != parameters.size() ) {
-            return FEELFnResult.ofError(new InvalidInputEvent(Severity.ERROR, "Illegal invocation of function", getName(), getName() + "( " + Arrays.asList(params)+" )", getSignature()));
-        }
-        
-        FEELEvent capturedException = null;
-        try {
-            ctx.enterFrame();
-            for ( int i = 0; i < parameters.size(); i++ ) {
-                final String paramName = parameters.get(i).name;
-                if (parameters.get(i).type.isAssignableValue(params[i])) {
-                    ctx.setValue(paramName, params[i]);
-                } else {
-                    ctx.setValue(paramName, null);
-                    ctx.notifyEvt(() -> {
-                        InvalidParametersEvent evt = new InvalidParametersEvent(Severity.WARN, paramName, "not conformant");
-                        evt.setNodeName(getName());
-                        evt.setActualParameters(parameters.stream().map(BaseFEELFunction.Param::getName).collect(Collectors.toList()),
-                                                Arrays.asList(params));
-                        return evt;
-                    });
-                }
-            }
-            Object result = this.body.evaluate( ctx );
-            return FEELFnResult.ofResult( result );
-        } catch( Exception e ) {
-            capturedException = new FEELEventBase(Severity.ERROR, "Error invoking function", new RuntimeException("Error invoking function " + getSignature() + ".", e));
-        } finally {
-            ctx.exitFrame();
-        }
-        return FEELFnResult.ofError( capturedException );
-    }
-
-    private String getSignature() {
-        return getName() + "( " + parameters.stream().map(p -> p.name + " : " + p.type).collect(Collectors.joining(", ")) + " )";
+    public CustomFEELFunction(String name, List<Param> parameters, BaseNode body) {
+        super(name, parameters, body);
     }
 
     @Override
-    protected boolean isCustomFunction() {
-        return true;
+    protected Object internalInvoke(EvaluationContext ctx) {
+        return this.body.evaluate(ctx);
     }
 
-    @Override
-    public String toString() {
-        return "function "+getSignature();
-    }
 }
